add toolbar entry + apply color mapping
