+++ src/renderer/HistogramCellRenderer.ts
@@ -266,7 +266,7 @@ export function getHistDOMRenderer(totalNumberOfRows: number, col: INumberColumn
       d.title = `${DEFAULT_FORMATTER(x0)} - ${DEFAULT_FORMATTER(x1)} (${length})`;
       d.dataset.x = DEFAULT_FORMATTER(x0);
       inner.style.height = `${Math.round(length * 100 / max)}%`;
-      inner.style.backgroundColor = colorOf(col, null, imposer);
+      inner.style.backgroundColor = colorOf(col, null, imposer, (x1 + x0) / 2);
     });
   };
   return {
